@@ -3,10 +3,10 @@
   "version": "1.1.0",
   "description": "Make a diff between two directories and list all missing files",
   "main": "cli.js",
-  "dependencies": {},
+  "type": "module",
   "devDependencies": {
-    "eslint": "^7.15.0",
-    "eslint-config-xo": "^0.33.1",
+    "eslint": "^8.11.0",
+    "eslint-config-xo": "^0.40.0",
     "faucet": "0.0.1",
     "tape": "^5.0.1"
   },
