@@ -1,12 +1,12 @@
-const test = require('tape')
-const diff = require('./diff')
+import test from 'tape'
+import diff from './diff.js'
 
 test('diff files, when using same tree', async assert => {
   const message = 'diff result must be empty'
 
   const actual = await diff({
     source: './fixtures/source',
-    target: './fixtures/source'
+    target: './fixtures/source',
   })
   const expected = []
 
@@ -20,7 +20,7 @@ test('diff files, when using distinct tree', async assert => {
 
   const actual = await diff({
     source: './fixtures/source',
-    target: './fixtures/target'
+    target: './fixtures/target',
   })
   const expected = []
 
@@ -34,7 +34,7 @@ test('diff files, when using tree with missing files', async assert => {
 
   const actual = await diff({
     source: './fixtures/source',
-    target: './fixtures/target/a'
+    target: './fixtures/target/a',
   })
   const expected = [
     '00.txt',
@@ -44,7 +44,7 @@ test('diff files, when using tree with missing files', async assert => {
     '06.txt',
     '07.txt',
     '08.txt',
-    '09.txt'
+    '09.txt',
   ]
 
   assert.deepEqual(actual, expected, message)
@@ -58,7 +58,7 @@ test('diff files, when using invalid source directory', async assert => {
   try {
     await diff({
       source: './fixtures/invalid',
-      target: './fixtures/target'
+      target: './fixtures/target',
     })
 
     assert.fail(message)
@@ -78,7 +78,7 @@ test('diff files, when using invalid target directory', async assert => {
   try {
     await diff({
       source: './fixtures/source',
-      target: './fixtures/invalid'
+      target: './fixtures/invalid',
     })
 
     assert.fail(message)
@@ -98,7 +98,7 @@ test('diff files, when using file as source directory', async assert => {
   try {
     await diff({
       source: './fixtures/file.txt',
-      target: './fixtures/target'
+      target: './fixtures/target',
     })
 
     assert.fail(message)
@@ -118,7 +118,7 @@ test('diff files, when using file as target directory', async assert => {
   try {
     await diff({
       source: './fixtures/source',
-      target: './fixtures/file.txt'
+      target: './fixtures/file.txt',
     })
 
     assert.fail(message)
