@@ -1,5 +1,5 @@
-const test = require('tape')
-const list = require('./list')
+import test from 'tape'
+import list from './list.js'
 
 test('list files, when using flat tree', async assert => {
   const message = 'should list 10 files'
@@ -15,7 +15,7 @@ test('list files, when using flat tree', async assert => {
     {path: 'fixtures/source', name: '06.txt'},
     {path: 'fixtures/source', name: '07.txt'},
     {path: 'fixtures/source', name: '08.txt'},
-    {path: 'fixtures/source', name: '09.txt'}
+    {path: 'fixtures/source', name: '09.txt'},
   ]
 
   assert.deepEqual(actual, expected, message)
@@ -37,7 +37,7 @@ test('list files, when using tree with branches', async assert => {
     {path: 'fixtures/target/b', name: '03.txt'},
     {path: 'fixtures/target/b/b', name: '04.txt'},
     {path: 'fixtures/target/c', name: '05.txt'},
-    {path: 'fixtures/target/c/c', name: '06.txt'}
+    {path: 'fixtures/target/c/c', name: '06.txt'},
   ]
 
   assert.deepEqual(actual, expected, message)
